Prevent hung process/threads when server_endpoint is not reachable
This commit is to address an issue that can cause Shaker to get
stuck indefinately waiting for a zeroMQ message that will never
arrive.

To reproduce set server_endpoint to 10.0.0.1:8080 (assuming
nothing is listening and/or this address is unreachable).
Shaker will go through launching the heartbeat thread, and
creating the heat stack, but then will be stuck at:
Waiting for quorum of agents:
If debug is turned on, there are no heartbeats coming through.

In order to fix this situation this commit sends 3 "tcp pings"
to the server_endpoint in order to test it's response before
the heartbeat thread and heat stack are created.

Change-Id: Ifcfacfb383e7553b53c5b2a20ad39ceccb833c8f